requirements.txt
quartz
quartz_discord
discord-ext-ipc
python-dotenv